chore: Update npm dependencies for Stripe and react-native-maps
This commit updates the npm dependencies for the Stripe library and the react-native-maps package. The Stripe library is updated from version 0.38.4 to 0.37.3, and the react-native-maps package is updated from version 1.18.0 to 1.14.0. This update ensures compatibility with the latest versions of these libraries and improves the overall stability and performance of the application.